src/requirements.txt
async-timeout==4.0.2
boto3==1.24.7
botocore==1.27.7
bottle==0.12.21
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Deprecated==1.2.13
dict2xml==1.7.1
flake8==4.0.1
idna==3.3
jmespath==1.0.0
jwt==1.3.1
mccabe==0.6.1
numpy==1.22.4
packaging==21.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.9
wrapt==1.14.1
xmltodict==0.13.0
